fix(pegboard): move sqlite v1 migration into actor workflow
