+++ engine/packages/pegboard/src/workflows/actor2/runtime.rs
@@ -372,7 +372,6 @@ pub async fn send_outbound(ctx: &ActivityCtx, input: &SendOutboundInput) -> Resu
 				// populated before it reaches the runner
 				hibernating_requests: Vec::new(),
 				preloaded_kv: None,
-				sqlite_schema_version: super::SQLITE_SCHEMA_VERSION_V2,
 				sqlite_startup_data: None,
 			});
