+++ engine/sdks/rust/envoy-client/src/config.rs
@@ -102,7 +102,6 @@ pub trait EnvoyCallbacks: Send + Sync + 'static {
 		generation: u32,
 		config: protocol::ActorConfig,
 		preloaded_kv: Option<protocol::PreloadedKv>,
-		sqlite_schema_version: u32,
 		sqlite_startup_data: Option<protocol::SqliteStartupData>,
 	) -> BoxFuture<anyhow::Result<()>>;
