+++ engine/sdks/rust/test-envoy/src/behaviors.rs
@@ -22,7 +22,6 @@ impl EnvoyCallbacks for DefaultTestCallbacks {
 		generation: u32,
 		_config: protocol::ActorConfig,
 		_preloaded_kv: Option<protocol::PreloadedKv>,
-		_sqlite_schema_version: u32,
 		_sqlite_startup_data: Option<protocol::SqliteStartupData>,
 	) -> BoxFuture<anyhow::Result<()>> {
 		Box::pin(async move {
