+++ rivetkit-rust/packages/rivetkit-core/tests/modules/task.rs
@@ -178,7 +178,6 @@ mod moved_tests {
 			_generation: u32,
 			_config: protocol::ActorConfig,
 			_preloaded_kv: Option<protocol::PreloadedKv>,
-			_sqlite_schema_version: u32,
 			_sqlite_startup_data: Option<protocol::SqliteStartupData>,
 		) -> EnvoyBoxFuture<anyhow::Result<()>> {
 			Box::pin(async { Ok(()) })
