+++ rivetkit-rust/packages/rivetkit/tests/client.rs
@@ -177,7 +177,6 @@ impl EnvoyCallbacks for IdleEnvoyCallbacks {
 		_generation: u32,
 		_config: protocol::ActorConfig,
 		_preloaded_kv: Option<protocol::PreloadedKv>,
-		_sqlite_schema_version: u32,
 		_sqlite_startup_data: Option<protocol::SqliteStartupData>,
 	) -> BoxFuture<anyhow::Result<()>> {
 		Box::pin(async { Ok(()) })
